+++ src/main/java/com/proyectTest/proyectTest/repository/DentistRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface DentistRepository extends JpaRepository<Dentist, Integer> {
+public interface DentistRepository extends JpaRepository<Dentist, Long> {
 }
